+++ src/library/DIPS.Mobile.UI/Components/Pickers/ItemPicker/ItemPicker.Mode.ContextMenu.cs
@@ -1,3 +1,4 @@
+using DIPS.Mobile.UI.API.Vibration;
 using DIPS.Mobile.UI.Components.ContextMenus;
 using DIPS.Mobile.UI.Extensions;
 
@@ -63,6 +64,7 @@ namespace DIPS.Mobile.UI.Components.Pickers.ItemPicker
 
         private void SetSelectedItemBasedOnContextMenuItem(ContextMenuItem item)
         {
+            VibrationService.SelectionChanged();
             SelectedItem = GetItemFromDisplayProperty(item.Title!);
         }
     }