+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [17.7.0] 
+- [ItemPicker] Added haptics for when people select a item from ItemPicker with context menu mode.
+- [SortControl] Added haptics for when people select a sort option.
+- [SegmentedControl] Added haptics for when people select/deselect a segment from the segmented control.
+
 ## [17.6.0] 
 - Resources was updated from DIPS.Mobile.DesignTokens